requirements.txt
flake8==6.1.0
pycodestyle==2.11.0
pyflakes==3.1.0
python-dotenv==1.0.0
requests==2.31.0
